Pre-commit and tests fixed
